pyproject.toml
[project]
name = "multi-agent-semantic-alignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gdown>=5.2.0",
    "lightning>=2.5.0.post0",
    "polars>=1.22.0",
    "python-dotenv>=1.0.1",
    "scipy>=1.15.1",
    "seaborn>=0.13.2",
    "torch>=2.6.0",
]
